refactor(065): move InitialPuzzle into data layer
