+++ app/src/test/java/org/cescfe/numpairs/ui/screen/GameViewModelStripEntryTest.kt
@@ -3,7 +3,7 @@ package org.cescfe.numpairs.ui.screen
 import org.cescfe.numpairs.domain.puzzle.Strip
 import org.cescfe.numpairs.domain.puzzle.StripEntryRange
 import org.cescfe.numpairs.domain.puzzle.StripItem
-import org.cescfe.numpairs.initialPuzzle
+import org.cescfe.numpairs.data.puzzle.seed.initialPuzzle
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Test
